fix(federated): review round 1 — empty-tier engine contract, honest anchor note, size splits (#184)
Empty PreferredTiers no longer short-circuits to Postgres at the engine
gate: it flows into EvaluateRoutingPolicy whose default decision carries
all three tiers, so direct engine callers get the same default-all
contract the service layer normalizes to (only PreferHot/[hot] gate now).
Pinned at both levels: a unit routing regression and a direct
env.Engine().Query e2e scenario that bypasses harness normalization.

The UseMainAsAnchor hint is recorded truthfully: the duckdb source no
longer reports it as PredicatePushdown (real pushdown facts live on the
pushdown-fragment source) and the note says "requested (advanced template
does not apply it)" instead of "honored".

Standards: tier_hints e2e split into scenarios + helpers files;
duckdb_query.go pure-move split (buildDuckDBQueryWithPlan /
serveFromPlanCache / duckDBParquetPathsForQuery -> duckdb_query_build.go,
the #220 seam) and buildDuckDBQueryWithPlan reduced to 82 lines via
buildDuckDBTemplateBaseParams extraction. All files/functions back under
the 500/100 caps.
